% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotChromPeakDensity-methods.R
\name{.compute_peak_density}
\alias{.compute_peak_density}
\title{Helper function to compute density and find feature groups}
\usage{
.compute_peak_density(pks, param)
}
\description{
Helper function to compute density and find feature groups
}
\keyword{internal}
