% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotChromPeakDensity-methods.R
\name{.descendMin}
\alias{.descendMin}
\title{Internal implementation of descendMin from XCMS
Finds the range around a maximum where density descends to local minima}
\usage{
.descendMin(y, istart = which.max(y))
}
\description{
Internal implementation of descendMin from XCMS
Finds the range around a maximum where density descends to local minima
}
\keyword{internal}
