% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_sample_data}
\alias{.get_sample_data}
\title{Get sample data from XCMS object}
\usage{
.get_sample_data(object)
}
\arguments{
\item{object}{XCMSnExp or XcmsExperiment object}
}
\value{
data.frame with sample metadata
}
\description{
Internal helper to handle both XCMSnExp and XcmsExperiment objects
}
\keyword{internal}
