% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_spectra_data}
\alias{.get_spectra_data}
\title{Get spectra data from XCMS object}
\usage{
.get_spectra_data(object)
}
\arguments{
\item{object}{XCMSnExp or XcmsExperiment object}
}
\value{
data.frame with spectra data including retention times
}
\description{
Internal helper to extract spectra/feature data from both XCMSnExp and XcmsExperiment objects
}
\keyword{internal}
