% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotChromatogramsOverlay-methods.R
\name{.gplotChromatogramsOverlay_impl}
\alias{.gplotChromatogramsOverlay_impl}
\title{Shared implementation function for gplotChromatogramsOverlay}
\usage{
.gplotChromatogramsOverlay_impl(
  object,
  col = "#00000060",
  type = "l",
  main = NULL,
  xlab = "retention time",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  peakType = c("polygon", "point", "rectangle", "none"),
  peakBg = NULL,
  peakCol = NULL,
  peakPch = 1,
  stacked = 0,
  transform = identity,
  ...
)
}
\description{
Shared implementation function for gplotChromatogramsOverlay
}
\keyword{internal}
