% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotChromPeakDensity-methods.R
\name{.gplotChromPeakDensity_impl}
\alias{.gplotChromPeakDensity_impl}
\title{Shared implementation function for gplotChromPeakDensity}
\usage{
.gplotChromPeakDensity_impl(
  object,
  param,
  col = "#00000060",
  xlab = "retention time",
  main = NULL,
  peakType = c("polygon", "point", "rectangle", "none"),
  peakCol = "#00000060",
  peakBg = "#00000020",
  peakPch = 1,
  simulate = TRUE,
  ...
)
}
\description{
Shared implementation function for gplotChromPeakDensity
}
\keyword{internal}
