% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotChromPeakDensity-methods.R
\name{.simulate_feature_groups}
\alias{.simulate_feature_groups}
\title{Helper function to simulate feature grouping based on peak density}
\usage{
.simulate_feature_groups(pks, dens_x, dens_y, param)
}
\description{
Helper function to simulate feature grouping based on peak density
}
\keyword{internal}
