% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_xcms_object}
\alias{.validate_xcms_object}
\title{Validate XCMS object type}
\usage{
.validate_xcms_object(object)
}
\arguments{
\item{object}{Object to validate}
}
\value{
TRUE if valid, stops with error otherwise
}
\description{
Validate XCMS object type
}
\keyword{internal}
