% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/gplotPrecursorIons-methods.R
\name{gplotPrecursorIons}
\alias{gplotPrecursorIons}
\alias{gplotPrecursorIons,MsExperiment-method}
\title{\emph{ggplot2} Version of \code{plotPrecursorIons()}}
\usage{
gplotPrecursorIons(object, pch = 21, col = "#00000080", bg = "#00000020", ...)

\S4method{gplotPrecursorIons}{MsExperiment}(object, pch = 21, col = "#00000080", bg = "#00000020", ...)
}
\arguments{
\item{object}{An \code{MsExperiment} object containing MS/MS data.}

\item{pch}{Point shape for precursor ions (default: \code{21} = filled circle).}

\item{col}{Point color (default: semi-transparent black).}

\item{bg}{Point background/fill color (default: very transparent black).}

\item{...}{Additional arguments passed to \emph{ggplot2} functions.}
}
\value{
A \code{ggplot} object (or list of \code{ggplot} objects if multiple files).
Use \code{+ labs()} to customize axis labels and titles.
}
\description{
Creates a \emph{ggplot2} version of precursor ion visualization for
\code{MsExperiment} objects. This function plots the m/z and retention time of
all precursor ions in MS2 spectra, useful for visualizing DDA
(Data-Dependent Acquisition) data.
}
\details{
This function visualizes the precursor ions selected for fragmentation in
MS/MS experiments. Each point represents a precursor ion, with:
\itemize{
\item X-axis: Retention time of the MS2 spectrum
\item Y-axis: Precursor m/z value
}

For \code{MsExperiment} objects with multiple files, separate plots are created
for each file.

The plot range includes all MS1 data to provide context, but only shows
precursor ions from MS2 spectra.

Default labels are provided (\code{"retention time"}, \code{"m/z"}), but can be
customized using \emph{ggplot2}'s \code{labs()} function, e.g.,
\code{gplotPrecursorIons(x) + labs(x = "RT (s)")}.
}
\examples{
library(xcmsVis)
library(MsExperiment)
library(ggplot2)
library(MsDataHub)

## Load a test data file with DDA LC-MS/MS data
fl <- MsDataHub::PestMix1_DDA.mzML()
pest_dda <- readMsExperiment(fl)

gplotPrecursorIons(pest_dda)

## Customize labels with ggplot2
gplotPrecursorIons(pest_dda) + labs(x = "RT (s)", y = "Precursor m/z",
    title = "DDA Analysis")

## Subset the data object to plot the data specifically for one or
## selected file/sample:
gplotPrecursorIons(pest_dda[1L])

}
\seealso{
\code{\link[xcms:plotPrecursorIons]{xcms::plotPrecursorIons()}} for the original \emph{xcms} implementation.
}
\author{
Jan Stanstrup
}
