% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms-utils.R
\name{xcms-utils}
\alias{xcms-utils}
\title{Internal XCMS utility functions}
\source{
https://github.com/sneumann/xcms/blob/devel/R/c.R

https://github.com/sneumann/xcms/blob/devel/R/do_adjustRtime-functions.R

https://github.com/sneumann/xcms/blob/devel/src/util.c
}
\description{
These functions are copied from the XCMS package to avoid using \code{:::} calls,
which would generate R CMD check NOTEs and fail Bioconductor standards.
}
\section{Why Copy Instead of Import}{

R packages cannot use \verb{package:::function()} to access internal functions
without generating a NOTE in R CMD check. Since Bioconductor requires
0 errors, 0 warnings, and 0 notes, we copy these stable utility functions
rather than depending on XCMS internals.
}

\section{Maintenance}{

\strong{Last synchronized:} 2024-11-17

\strong{XCMS source version:} devel branch (accessed 2024-11-17)

\strong{XCMS GitHub:} https://github.com/sneumann/xcms

\strong{Functions copied:}
\itemize{
\item \code{.descendMin()} - C + R wrapper from \code{R/c.R} and \code{src/util.c}
\item \code{.applyRtAdjustment()} - Pure R from \code{R/do_adjustRtime-functions.R}
\item \code{.rt_model()} - Pure R from \code{R/do_adjustRtime-functions.R}
\item \code{.check_gam_library()} - Helper for \code{.rt_model()}
}
}

\section{Update Process}{

To update these functions:
\enumerate{
\item Check XCMS source at https://github.com/sneumann/xcms/tree/devel
\item Compare with current implementation in this file
\item Update both R wrapper and C code (if changed)
\item Run tests in \code{tests/testthat/test-xcms-utils.R}
\item Update "Last synchronized" date above
}
}

\section{Stability}{

These are low-level utility functions that have remained stable in XCMS
for many years. Breaking changes are unlikely but should be monitored.
}

\section{Alternative Approach}{

If these functions are needed by multiple packages, consider requesting
that XCMS maintainers export them officially. Open an issue at:
https://github.com/sneumann/xcms/issues
}

\author{
Original XCMS authors (Colin A. Smith, Ralf Tautenhahn, Steffen Neumann,
Johannes Rainer, et al.)
}
\keyword{internal}
